Tmux: Add template option for status-right
This allows for smarter section lazy loading if I want to.
By providing an estimate of the length I can delay executing the command that
runs to produce the section contents.
Before this the first hidden section would always be evaluated, but now it can
be stopped before it is evaluated due to the estimate.
If there is not sufficient room for the estimate then the section is skipped
without evaluating the template for the content.
The final content is still checked for length as before.
This change should be essentially backwards compatible.